Refactored code to use (en|de)codeInt funcs
Introduced functions which take and return *big.Int instead of plain
strings, anticipating the introduction of other
inputs/outputs. Basically, the code converts, at some point, to
big.Int to avoid reinventing the wheel (which is good) so this is
probably our best "common base format" to call the core
algorithm. It's not exposed in public funcs as using big.Int is, to
some extent, an implementation detail.