Optimized code by puting map calc in init() func
Same as for base calculation, pre-calculating the dictionnary at
program startup should make things faster. Note that it's going to
"cost" some resident memory as a map will be stored when module is
imported, even if decode is never used, but this should be better than
creating it / destroying it all the time. Garbage collection does have
a cost too. Did speed up "Decode()" by a factor 3 on my laptop.